# projects/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "projects-service dev",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/project_status": {
      "get": {
        "tags": [
          "ProjectStatus"
        ],
        "summary": "Get Project Statuses",
        "description": "Gets a paginated list of project statuses",
        "operationId": "get_project_statuses_api_project_status_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1.0,
              "type": "integer",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100.0,
              "minimum": 1.0,
              "type": "integer",
              "default": 50
            },
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponsePaginated_IProjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProjectStatus"
        ],
        "summary": "Create Project Status",
        "description": "Creates a new project status",
        "operationId": "create_project_status_api_project_status_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IProjectStatusCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPostResponseBase_IProjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/project_status/{id}": {
      "get": {
        "tags": [
          "ProjectStatus"
        ],
        "summary": "Read Project Status By Id",
        "description": "Gets a project status by id",
        "operationId": "read_project_status_by_id_api_project_status__id__get",
        "parameters": [
          {
            "description": "The UUID id of the project status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IProjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProjectStatus"
        ],
        "summary": "Update Project Status",
        "description": "Updates a project status by its id",
        "operationId": "update_project_status_api_project_status__id__put",
        "parameters": [
          {
            "description": "The UUID id of the project status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IProjectStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPutResponseBase_IProjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProjectStatus"
        ],
        "summary": "Delete Project Status",
        "description": "Deletes a project status by its id",
        "operationId": "delete_project_status_api_project_status__id__delete",
        "parameters": [
          {
            "description": "The UUID id of the project status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IDeleteResponseBase_IProjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/project_status/{id}/projects": {
      "get": {
        "tags": [
          "ProjectStatus"
        ],
        "summary": "Get Project Status With Projects List By Id",
        "description": "Gets a project status with projects list by id",
        "operationId": "get_project_status_with_projects_list_by_id_api_project_status__id__projects_get",
        "parameters": [
          {
            "description": "The UUID id of the project status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IProjectStatusWithProjectsRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/project_status/{id}/subprojects": {
      "get": {
        "tags": [
          "ProjectStatus"
        ],
        "summary": "Get Project Status With Subprojects List By Id",
        "description": "Gets a project status with subprojects list by id",
        "operationId": "get_project_status_with_subprojects_list_by_id_api_project_status__id__subprojects_get",
        "parameters": [
          {
            "description": "The UUID id of the project status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IProjectStatusWithSubrojectsRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/project_status/{id}/full": {
      "get": {
        "tags": [
          "ProjectStatus"
        ],
        "summary": "Get Project Status With Subprojects And Projects List By Id",
        "description": "Gets a project status with subprojects and projects list by id",
        "operationId": "get_project_status_with_subprojects_and_projects_list_by_id_api_project_status__id__full_get",
        "parameters": [
          {
            "description": "The UUID id of the project status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IProjectStatusWithSubrojectsAndProjectsRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/project/list/status": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get Projects List By Project Status Id",
        "description": "Gets a paginated list of projects by status id",
        "operationId": "get_projects_list_by_project_status_id_api_project_list_status_get",
        "parameters": [
          {
            "description": "The UUID id of the project status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project status",
              "format": "uuid"
            },
            "name": "id",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1.0,
              "type": "integer",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100.0,
              "minimum": 1.0,
              "type": "integer",
              "default": 50
            },
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponsePaginated_IProjectRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Read Project By Id",
        "description": "Gets a project by id",
        "operationId": "read_project_by_id_api_project__id__get",
        "parameters": [
          {
            "description": "The UUID id of the project",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IProjectRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "summary": "Update Project",
        "description": "Updates a project by its id",
        "operationId": "update_project_api_project__id__put",
        "parameters": [
          {
            "description": "The UUID id of the project",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IProjectUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPutResponseBase_IProjectRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Delete Project",
        "description": "Deletes a project by its id",
        "operationId": "delete_project_api_project__id__delete",
        "parameters": [
          {
            "description": "The UUID id of the project",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IDeleteResponseBase_IProjectRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{id}/subprojects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get Project With Subprojects List By Id",
        "description": "Gets a project by id",
        "operationId": "get_project_with_subprojects_list_by_id_api_project__id__subprojects_get",
        "parameters": [
          {
            "description": "The UUID id of the project",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IProjectWithSubprojectsListRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{id}/status": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get Project With Status By Id",
        "description": "Gets a project by id",
        "operationId": "get_project_with_status_by_id_api_project__id__status_get",
        "parameters": [
          {
            "description": "The UUID id of the project",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IProjectWithStatus_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{id}/full": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get Full Project Info By Id",
        "description": "Gets a full project information by id",
        "operationId": "get_full_project_info_by_id_api_project__id__full_get",
        "parameters": [
          {
            "description": "The UUID id of the project",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IFullProjectInfoRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/project": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create Project",
        "description": "Creates a new project",
        "operationId": "create_project_api_project_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPostResponseBase_IProjectRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/subproject/list/project": {
      "get": {
        "tags": [
          "Subprojects"
        ],
        "summary": "Get Subprojects List By Project Id",
        "description": "Gets a paginated list of subprojects by project id",
        "operationId": "get_subprojects_list_by_project_id_api_subproject_list_project_get",
        "parameters": [
          {
            "description": "The UUID id of the project",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project",
              "format": "uuid"
            },
            "name": "id",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1.0,
              "type": "integer",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100.0,
              "minimum": 1.0,
              "type": "integer",
              "default": 50
            },
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponsePaginated_ISubprojectRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/subproject/{id}/status": {
      "get": {
        "tags": [
          "Subprojects"
        ],
        "summary": "Get Subproject With Status By Id",
        "description": "Gets a subproject with status by id",
        "operationId": "get_subproject_with_status_by_id_api_subproject__id__status_get",
        "parameters": [
          {
            "description": "The UUID id of the subproject",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the subproject",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_ISubprojectWithStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subproject/{id}/project": {
      "get": {
        "tags": [
          "Subprojects"
        ],
        "summary": "Get Subproject With Project By Id",
        "description": "Gets a subproject with project by id",
        "operationId": "get_subproject_with_project_by_id_api_subproject__id__project_get",
        "parameters": [
          {
            "description": "The UUID id of the subproject",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the subproject",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_ISubprojectWithProjectRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tower/list/status": {
      "get": {
        "tags": [
          "Towers"
        ],
        "summary": "Get Towers List By Status Id",
        "description": "Gets a paginated list of towers by status id",
        "operationId": "get_towers_list_by_status_id_api_tower_list_status_get",
        "parameters": [
          {
            "description": "The UUID id of the object status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the object status",
              "format": "uuid"
            },
            "name": "id",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1.0,
              "type": "integer",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100.0,
              "minimum": 1.0,
              "type": "integer",
              "default": 50
            },
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponsePaginated_ITowerRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tower/{id}/status": {
      "get": {
        "tags": [
          "Towers"
        ],
        "summary": "Get Tower With Object Status By Id",
        "description": "Gets a tower with status by id",
        "operationId": "get_tower_with_object_status_by_id_api_tower__id__status_get",
        "parameters": [
          {
            "description": "The UUID id of the tower",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the tower",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_ITowerWithObjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "IGetResponsePaginated_IProjectStatusRead_": {
        "title": "IGetResponsePaginated[IProjectStatusRead]",
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/PageBase_DataType__IProjectStatusRead_"
          }
        }
      },
...
      "IPostResponseBase_IProjectStatusRead_": {
        "title": "IPostResponseBase[IProjectStatusRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data created correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/IProjectStatusRead"
          }
        }
      },
...
      "IProjectStatusCreate": {
        "title": "IProjectStatusCreate",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        }
      },
...
    }
  }
}